requirements.txt
streamlit
google-generativeai
langchain-google-genai
langchain
langchain-chroma
sentence-transformers
pypdf
python-dotenv
langchain_community
